requirements.txt
anyio==3.5.0
asgiref==3.5.1
bcrypt==3.2.2
cffi==1.15.0
click==8.1.3
cryptography==37.0.1
fastapi==0.75.2
FastAPI-SQLAlchemy==0.2.1
greenlet==1.1.2
h11==0.13.0
idna==3.3
psycopg2-binary==2.9.3
pycparser==2.21
pydantic==1.9.0
python-http-client==3.3.7
sendgrid==6.9.7
sniffio==1.2.0
SQLAlchemy==1.4.36
starkbank-ecdsa==2.0.3
starlette==0.17.1
typing-extensions==4.2.0
uvicorn==0.17.6
gunicorn
pyjwt==2.3.0
